add common jinja exceptions
