+++ core/dbt/common/exceptions/__init__.py
@@ -3,3 +3,5 @@ from dbt.common.exceptions.events import *  # noqa
 from dbt.common.exceptions.macros import *  # noqa
 from dbt.common.exceptions.contracts import *  # noqa
 from dbt.common.exceptions.connection import *  # noqa
+from dbt.common.exceptions.system import *  # noqa
+from dbt.common.exceptions.jinja import *  # noqa
