+++ core/dbt/deps/git.py
@@ -10,7 +10,8 @@ from dbt.contracts.project import (
     GitPackage,
 )
 from dbt.deps.base import PinnedPackage, UnpinnedPackage, get_downloads_path
-from dbt.exceptions import ExecutableError, MultipleVersionGitDepsError
+from dbt.common.exceptions import ExecutableError
+from dbt.exceptions import MultipleVersionGitDepsError
 from dbt.common.events.functions import fire_event, warn_or_error, scrub_secrets, env_secrets
 from dbt.events.types import EnsureGitInstalled, DepsUnpinned, DepsScrubbedPackageName
 from dbt.utils import md5
