+++ core/dbt/exceptions.py
@@ -12,7 +12,7 @@ from dbt.common.exceptions import (
     env_secrets,
     scrub_secrets,
     DbtValidationError,
-    CommandError,
+    CommandResultError,
 )
 from dbt.node_types import NodeType, AccessType
 
@@ -136,39 +136,6 @@ class DbtProfileError(DbtConfigError):
     pass
 
 
-class ExecutableError(CommandError):
-    def __init__(self, cwd: str, cmd: List[str], msg: str) -> None:
-        super().__init__(cwd, cmd, msg)
-
-
-class WorkingDirectoryError(CommandError):
-    def __init__(self, cwd: str, cmd: List[str], msg: str) -> None:
-        super().__init__(cwd, cmd, msg)
-
-    def __str__(self):
-        return f'{self.msg}: "{self.cwd}"'
-
-
-class CommandResultError(CommandError):
-    def __init__(
-        self,
-        cwd: str,
-        cmd: List[str],
-        returncode: Union[int, Any],
-        stdout: bytes,
-        stderr: bytes,
-        msg: str = "Got a non-zero returncode",
-    ) -> None:
-        super().__init__(cwd, cmd, msg)
-        self.returncode = returncode
-        self.stdout = scrub_secrets(stdout.decode("utf-8"), env_secrets())
-        self.stderr = scrub_secrets(stderr.decode("utf-8"), env_secrets())
-        self.args = (cwd, self.cmd, returncode, self.stdout, self.stderr, msg)
-
-    def __str__(self):
-        return f"{self.msg} running: {self.cmd}"
-
-
 class InvalidSelectorError(DbtRuntimeError):
     def __init__(self, name: str) -> None:
         self.name = name
@@ -208,49 +175,6 @@ class MaterializtionMacroNotUsedError(CompilationError):
         super().__init__(msg=self.msg)
 
 
-class MissingControlFlowStartTagError(CompilationError):
-    def __init__(self, tag, expected_tag: str, tag_parser) -> None:
-        self.tag = tag
-        self.expected_tag = expected_tag
-        self.tag_parser = tag_parser
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        linepos = self.tag_parser.linepos(self.tag.start)
-        msg = (
-            f"Got an unexpected control flow end tag, got {self.tag.block_type_name} but "
-            f"expected {self.expected_tag} next (@ {linepos})"
-        )
-        return msg
-
-
-class UnexpectedControlFlowEndTagError(CompilationError):
-    def __init__(self, tag, expected_tag: str, tag_parser) -> None:
-        self.tag = tag
-        self.expected_tag = expected_tag
-        self.tag_parser = tag_parser
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        linepos = self.tag_parser.linepos(self.tag.start)
-        msg = (
-            f"Got an unexpected control flow end tag, got {self.tag.block_type_name} but "
-            f"never saw a preceeding {self.expected_tag} (@ {linepos})"
-        )
-        return msg
-
-
-class UnexpectedMacroEOFError(CompilationError):
-    def __init__(self, expected_name: str, actual_name: str) -> None:
-        self.expected_name = expected_name
-        self.actual_name = actual_name
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        msg = f'unexpected EOF, expected {self.expected_name}, got "{self.actual_name}"'
-        return msg
-
-
 class MacroNamespaceNotStringError(CompilationError):
     def __init__(self, kwarg_type: Any) -> None:
         self.kwarg_type = kwarg_type
@@ -261,47 +185,6 @@ class MacroNamespaceNotStringError(CompilationError):
             "The macro_namespace parameter to adapter.dispatch "
             f"is a {self.kwarg_type}, not a string"
         )
-        return msg
-
-
-class NestedTagsError(CompilationError):
-    def __init__(self, outer, inner) -> None:
-        self.outer = outer
-        self.inner = inner
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        msg = (
-            f"Got nested tags: {self.outer.block_type_name} (started at {self.outer.start}) did "
-            f"not have a matching {{{{% end{self.outer.block_type_name} %}}}} before a "
-            f"subsequent {self.inner.block_type_name} was found (started at {self.inner.start})"
-        )
-        return msg
-
-
-class BlockDefinitionNotAtTopError(CompilationError):
-    def __init__(self, tag_parser, tag_start) -> None:
-        self.tag_parser = tag_parser
-        self.tag_start = tag_start
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        position = self.tag_parser.linepos(self.tag_start)
-        msg = (
-            f"Got a block definition inside control flow at {position}. "
-            "All dbt block definitions must be at the top level"
-        )
-        return msg
-
-
-class MissingCloseTagError(CompilationError):
-    def __init__(self, block_type_name: str, linecount: int) -> None:
-        self.block_type_name = block_type_name
-        self.linecount = linecount
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        msg = f"Reached EOF without finding a close tag for {self.block_type_name} (searched from line {self.linecount})"
         return msg
 
 
@@ -370,20 +253,6 @@ class OperationError(CompilationError):
     def get_message(self) -> str:
         msg = (
             f"dbt encountered an error when attempting to create a {self.operation_name}. "
-            "If this error persists, please create an issue at: \n\n"
-            "https://github.com/dbt-labs/dbt-core"
-        )
-
-        return msg
-
-
-class SymbolicLinkError(CompilationError):
-    def __init__(self) -> None:
-        super().__init__(msg=self.get_message())
-
-    def get_message(self) -> str:
-        msg = (
-            "dbt encountered an error when attempting to create a symbolic link. "
             "If this error persists, please create an issue at: \n\n"
             "https://github.com/dbt-labs/dbt-core"
         )
